refactor: rename process instance modification handler
